Mediculus Fall and Pain Detection Model/requirements.txt
boto==2.49.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
future==0.18.2
idna==2.10
importlib-metadata==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.3.0.38
openpifpaf==0.12.14
pandas==1.3.4
Pillow==8.4.0
pyparsing==2.4.7
pysparkling==0.6.0
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2020.1
requests==2.24.0
scipy==1.7.1
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.25.11
zipp==3.6.0
